Add scout-off-chain job to collect data that is not collected from the chain
